folder renamed and files content changed
