+++ Dockerfile
@@ -20,4 +20,4 @@ RUN adduser -u 5678 --disabled-password --gecos "" appuser && chown -R appuser /
 USER appuser
 
 # During debugging, this entry point will be overridden. For more information, please refer to https://aka.ms/vscode-docker-python-debug
-CMD ["python", "rps_game.py"]
+CMD ["python", "RPS\rps_game.py"]
